requirements.txt
pandas
flask 
faiss-cpu 
sentence-transformers 
mysql-connector-python
